services/analyzer/requirements.txt
# services/analyzer/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.33.0
torch==2.5.1
transformers==4.47.1
accelerate==1.2.1
sentencepiece==0.2.0
protobuf==5.29.2
numpy==1.26.4
scikit-learn==1.6.0
